requirements.txt
# Core chemistry and mapping
rdkit-pypi>=2022.9.5
rxnmapper>=0.3.0,<0.5

# Data and utilities
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.26,<2
networkx>=3.0

# CLI and UX
click>=8.1.3
tqdm>=4.65.0
ujson>=5.8.0

# Testing
pytest>=7.3.0

# Optional utilities (install if desired)
# datamol>=0.10.0
# rdchiral>=1.1.0
